be/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aniso8601==7.0.0          # via graphene
appdirs==1.4.3            # via black
asgiref==3.2.5            # via django
attrs==19.3.0             # via black, flake8-bugbear, pytest
berglas-python==0.3.3     # via -r requirements.in
black==19.10b0            # via -r requirements.in, flake8-black
bleach==3.1.3             # via django-markdownify
cachetools==4.0.0         # via google-auth
certifi==2019.11.28       # via httpx, requests
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via httpx, requests
click==7.1.1              # via black, pip-tools
cryptography==2.8         # via berglas-python, social-auth-core
data==0.4                 # via latex
decorator==4.4.2          # via data
defusedxml==0.6.0         # via python3-openid, social-auth-core
dj-database-url==0.5.0    # via -r requirements.in
django-cors-headers==3.2.1  # via -r requirements.in
django-debug-toolbar-request-history==0.1.1  # via -r requirements.in
django-debug-toolbar==2.2  # via -r requirements.in, django-debug-toolbar-request-history
django-markdownify==0.8.0  # via -r requirements.in
django-storages[google]==1.9.1  # via -r requirements.in
django==3.0.4             # via -r requirements.in, django-cors-headers, django-debug-toolbar, django-markdownify, django-storages
ecdsa==0.15               # via python-jose
entrypoints==0.3          # via flake8
fancycompleter==0.9.1     # via pdbpp
flake8-black==0.1.1       # via -r requirements.in
flake8-bugbear==20.1.4    # via -r requirements.in
flake8==3.7.9             # via -r requirements.in, flake8-black, flake8-bugbear
funcsigs==1.0.2           # via data
future==0.18.2            # via latex
google-api-core[grpc]==1.16.0  # via google-cloud-core, google-cloud-kms, google-cloud-pubsub, google-cloud-secret-manager, google-cloud-tasks
google-auth==1.11.3       # via google-api-core, google-cloud-storage
google-cloud-core==1.3.0  # via google-cloud-storage
google-cloud-kms==1.3.0   # via berglas-python
google-cloud-pubsub==1.3.1  # via -r requirements.in
google-cloud-secret-manager==0.2.0  # via -r requirements.in
google-cloud-storage==1.26.0  # via -r requirements.in, berglas-python, django-storages
google-cloud-tasks==1.5.0  # via -r requirements.in
google-resumable-media==0.5.0  # via google-cloud-storage
googleapis-common-protos[grpc]==1.51.0  # via google-api-core, grpc-google-iam-v1
graphene==3.0.dev20190817210753  # via -r requirements.in
graphql-core==3.0.3       # via graphene, graphql-relay
graphql-relay==3.0.0      # via graphene
grpc-google-iam-v1==0.12.3  # via google-cloud-kms, google-cloud-pubsub, google-cloud-secret-manager, google-cloud-tasks
grpcio==1.27.2            # via google-api-core, googleapis-common-protos, grpc-google-iam-v1
gunicorn==20.0.4          # via -r requirements.in
h11==0.9.0                # via httpx
h2==3.2.0                 # via httpx
hpack==3.0.0              # via h2
hstspreload==2020.3.17    # via httpx
httpx==0.12.1             # via -r requirements.in
hyperframe==5.2.0         # via h2
idna==2.9                 # via httpx, requests
importlib-metadata==1.5.0  # via pluggy, pytest
latex==0.7.0              # via -r requirements.in
markdown==3.2.1           # via django-markdownify
mccabe==0.6.1             # via flake8
more-itertools==8.2.0     # via pytest
oauthlib==3.1.0           # via requests-oauthlib, social-auth-core
packaging==20.3           # via pytest
pathspec==0.7.0           # via black
pdbpp==0.10.2             # via -r requirements.in
phonenumbers==8.12.0      # via -r requirements.in
pip-tools==4.5.1          # via -r requirements.in
pluggy==0.13.1            # via pytest
protobuf==3.11.3          # via google-api-core, googleapis-common-protos
psycopg2-binary==2.8.4    # via -r requirements.in
py==1.8.1                 # via pytest
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, python-jose, rsa
pycodestyle==2.5.0        # via flake8
pycparser==2.20           # via cffi
pyflakes==2.1.1           # via flake8
pygments==2.6.1           # via pdbpp
pyjwt==1.7.1              # via -r requirements.in, social-auth-core, twilio
pyparsing==2.4.6          # via packaging
pyrepl==0.9.0             # via fancycompleter
pytest-django==3.8.0      # via -r requirements.in
pytest-dotenv==0.4.0      # via -r requirements.in
pytest==5.4.1             # via -r requirements.in, pytest-django, pytest-dotenv
python-dateutil==2.8.1    # via -r requirements.in
python-dotenv==0.12.0     # via -r requirements.in, pytest-dotenv
python-jose==3.1.0        # via -r requirements.in
python3-openid==3.1.0     # via social-auth-core
pytz==2019.3              # via django, google-api-core, twilio
regex==2020.2.20          # via -r requirements.in, black
requests-oauthlib==1.3.0  # via social-auth-core
requests==2.23.0          # via -r requirements.in, google-api-core, requests-oauthlib, social-auth-core, twilio
rfc3986==1.3.2            # via httpx
rsa==4.0                  # via google-auth, python-jose
shutilwhich==1.1.0        # via latex
six==1.14.0               # via bleach, cryptography, data, ecdsa, google-api-core, google-auth, google-resumable-media, grpcio, packaging, pip-tools, protobuf, python-dateutil, python-jose, social-auth-app-django, social-auth-core, twilio
sniffio==1.1.0            # via httpx
social-auth-app-django==3.1.0  # via -r requirements.in
social-auth-core==3.3.0   # via social-auth-app-django
sqlparse==0.3.1           # via django, django-debug-toolbar
tempdir==0.7.1            # via latex
toml==0.10.0              # via black
tqdm==4.43.0              # via -r requirements.in
twilio==6.37.0            # via -r requirements.in
typed-ast==1.4.1          # via black
ukpostcodeparser==1.1.2   # via -r requirements.in
unidecode==1.1.1          # via social-auth-core
urllib3==1.25.8           # via httpx, requests
wcwidth==0.1.8            # via pytest
webencodings==0.5.1       # via bleach
whitenoise==5.0.1         # via -r requirements.in
wmctrl==0.3               # via pdbpp
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
